Python 3 compatibility and use system sessions.
The code has been reworked to be compatible with Python 3, because Python 3
is reaching EOL.

Also, updated the authentication so it will be done only once, to obtain
a system session. Subsequent requests use the session id. This is needed
because some Graylog deployments will randonmly refuse connections because
of rate limits in the authentication backends.